working on routes

+++ routes/api/product-routes.js
@@ -3,44 +3,41 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // The `/api/products` endpoint
 
-// get all products
-router.get('/', (req, res) => {
-  // find all products
-  // be sure to include its associated Category and Tag data
-  try{
+// Get all products and associated Category and Tag data
+router.get('/', async (req, res) => {
+  try {
     const productData = await Product.findAll({
-      //Category and Tag tables
-      include: [{model: Category}, {model: Tag}],
+      include: [{ model: Category }]
+    }, {
+      include: [{ model: Tag, through: ProductTag, as: 'products_with_tags' }]
     });
-    res.status(200).json(productData)
-  } catch (err){
+    res.status(200).json(productData);
+  } catch (err) {
     res.status(500).json(err);
   }
 });
 
-// get one product
-router.get('/:id', (req, res) => {
-  // find a single product by its `id`
-  // be sure to include its associated Category and Tag data
-  try{
+// Get product by id with its associated Category and Tag data
+router.get('/:id', async (req, res) => {
+  try {
     const productData = await Product.findByPk(req.params.id, {
-      // join Category and Tag tables 
-      include: [{model: Category}, {model: Tag}],
-    });
+      include: [{ model: Category }]
+    }, {
+      include: [{ model: Tag }]
+    })
 
-    if(!productData){
-      res.status(404).json({ message: 'No product found with this id!'});
+    if (!productData) {
+      res.status(404).json({ message: 'No product with this id!' });
       return;
     }
 
     res.status(200).json(productData);
-  } catch (err){
+  } catch (err) {
     res.status(500).json(err);
   }
-    
 });
 
-// create new product
+// Create new product
 router.post('/', (req, res) => {
   /* req.body should look like this...
     {
@@ -52,8 +49,9 @@ router.post('/', (req, res) => {
   */
   Product.create(req.body)
     .then((product) => {
+      //if no tagids --> create error message
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      if (req.body.tagIds.length) {
+      if (req.body.tagIds?.length) {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
             product_id: product.id,
@@ -72,40 +70,45 @@ router.post('/', (req, res) => {
     });
 });
 
-// update product
 router.put('/:id', (req, res) => {
   // update product data
   Product.update(req.body, {
     where: {
       id: req.params.id,
-    },
+    }
   })
     .then((product) => {
+      // if no tagIds in req.body --> skip modification of tagids
+      if (!req.body.tagIds) {
+        return product;
+      }
       // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
+      return ProductTag.findAll({ where: { product_id: req.params.id } })
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
+        .then((productTags) => {
+          
+          // get list of current tag_ids
+          const productTagIds = productTags.map(({ tag_id }) => tag_id);
+          // create filtered list of new tag_ids
+          const newProductTags = req.body.tagIds
+            .filter((tag_id) => !productTagIds.includes(tag_id))
+            .map((tag_id) => {
+              return {
+                product_id: req.params.id,
+                tag_id,
+              };
+            });
+          // figure out which ones to remove
+          const productTagsToRemove = productTags
+            .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+            .map(({ id }) => id);
+
+          // run both actions
+          return Promise.all([
+            ProductTag.destroy({ where: { id: productTagsToRemove } }),
+            ProductTag.bulkCreate(newProductTags),
+          ]);
+        })
     })
     .then((updatedProductTags) => res.json(updatedProductTags))
     .catch((err) => {
@@ -114,7 +117,7 @@ router.put('/:id', (req, res) => {
     });
 });
 
-router.delete('/:id', (req, res) => {
+router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
   try {
     const productData = await Product.destroy({
@@ -123,15 +126,16 @@ router.delete('/:id', (req, res) => {
       }
     });
 
-    if(!productData){
-      res.status(404).json({ message: 'No product found with this id!'});
+    if (!productData) {
+      res.status(404).json({ message: 'No product found with this id!' });
       return;
     }
 
     res.status(200).json(productData);
-  } catch (err){
-    res.status(500).json(err)
+  } catch (err) {
+    res.status(500).json(err);
   }
 });
+
 
 module.exports = router;
